Move controller directory to services
