+++ src/main/java/de/fearnixx/jeak/service/controller/connection/IConnectionVerifier.java
@@ -1,4 +1,4 @@
-package de.fearnixx.jeak.controller.connection;
+package de.fearnixx.jeak.service.controller.connection;
 
 public interface IConnectionVerifier {
     boolean verifyRequest(Class<?> controllerClass, String authorizationText);
