+++ src/main/java/de/fearnixx/jeak/service/controller/reflect/RequestParam.java
@@ -1,4 +1,4 @@
-package de.fearnixx.jeak.controller.reflect;
+package de.fearnixx.jeak.service.controller.reflect;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
